requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
djoser>=2.2
drf-yasg>=1.21
python-dotenv>=1.0
gunicorn>=21.2
whitenoise>=6.6
openai>=1.30     # For quiz generation
Pillow
corsheaders           # For image handling (if avatars)
youtube-transcript-api
openai
google-generativeai

